# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fulcrum-India API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schemes:
    get:
      operationId: listSchemes
      tags: [schemes]
      summary: List government schemes
      parameters:
        - name: community
          in: query
          schema:
            type: string
        - name: sector
          in: query
          schema:
            type: string
        - name: stage
          in: query
          schema:
            type: string
        - name: minFunding
          in: query
          schema:
            type: integer
        - name: maxFunding
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of schemes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scheme"
  /schemes/{id}:
    get:
      operationId: getScheme
      tags: [schemes]
      summary: Get scheme by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scheme
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scheme"
  /schemes/matches/{entrepreneurId}:
    get:
      operationId: getSchemeMatches
      tags: [schemes]
      summary: Get scheme matches for an entrepreneur
      parameters:
        - name: entrepreneurId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Matched schemes with scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SchemeMatch"

  # DOCUMENTS
...
  /messages/{conversationId}:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
...
components:
  schemas:
...
    Scheme:
      type: object
      required: [id, name, ministry, minFunding, maxFunding, subsidyPercent, eligibleCommunities, eligibleSectors, avgProcessingDays, successRate, applicationsCount]
      properties:
        id:
          type: integer
        name:
          type: string
        ministry:
          type: string
        description:
          type: string
        minFunding:
          type: integer
        maxFunding:
          type: integer
        subsidyPercent:
          type: ["number", "null"]
        subsidyLabel:
          type: ["string", "null"]
        eligibleCommunities:
          type: array
          items:
            type: string
        eligibleSectors:
          type: array
          items:
            type: string
        avgProcessingDays:
          type: integer
        successRate:
          type: number
        applicationsCount:
          type: integer
        builderTip:
          type: ["string", "null"]
        builderTipAuthor:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
    SchemeMatch:
      type: object
      required: [scheme, matchScore, matchReasons]
      properties:
        scheme:
          $ref: "#/components/schemas/Scheme"
        matchScore:
          type: integer
        matchReasons:
          type: array
          items:
            type: string
        localStats:
          type: object
          properties:
            applications:
              type: integer
            approved:
              type: number
            avgDays:
              type: integer

    # DOCUMENT
...
    Message:
      type: object
      required: [id, conversationId, senderId, senderName, senderRole, content, sentAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        senderRole:
          type: string
        content:
          type: string
        sentAt:
          type: string
        isRead:
          type: boolean
        attachmentUrl:
          type: ["string", "null"]
        attachmentName:
          type: ["string", "null"]
...